+++ OgreMain/src/OgreInstanceBatchShader.cpp
@@ -198,7 +198,7 @@ namespace Ogre
             thisVertexData->vertexBufferBinding->setBinding( lastSource, vertexBuffer );
 
             char* thisBuf = static_cast<char*>(vertexBuffer->lock(HardwareBuffer::HBL_DISCARD));
-            for( size_t j=0; j<mInstancesPerBatch; ++j )
+            for( uint8 j=0; j<uint8(mInstancesPerBatch); ++j )
             {
                 for( size_t k=0; k<baseVertexData->vertexCount; ++k )
                 {
@@ -254,7 +254,7 @@ namespace Ogre
                     originalVal = *initBuf32++;
 
                 if( indexType == HardwareIndexBuffer::IT_16BIT )
-                    *thisBuf16++ = static_cast<uint16>(originalVal) + vertexOffset;
+                    *thisBuf16++ = static_cast<uint16>(originalVal + vertexOffset);
                 else
                     *thisBuf32++ = static_cast<uint32>(originalVal + vertexOffset);
             }
@@ -267,7 +267,7 @@ namespace Ogre
     void InstanceBatchShader::setupHardwareSkinned( const SubMesh* baseSubMesh, VertexData *thisVertexData,
                                                     VertexData *baseVertexData )
     {
-        const size_t numBones = baseSubMesh->blendIndexToBoneIndexMap.size();
+        const uint8 numBones = uint8(baseSubMesh->blendIndexToBoneIndexMap.size());
         mNumWorldMatrices = mInstancesPerBatch * numBones;
 
         for( uint16 i=0; i<=thisVertexData->vertexDeclaration->getMaxSource(); ++i )
@@ -292,7 +292,7 @@ namespace Ogre
             char *startBuf = baseBuf;
 
             //Copy and repeat
-            for( size_t j=0; j<mInstancesPerBatch; ++j )
+            for( uint8 j=0; j<uint8(mInstancesPerBatch); ++j )
             {
                 //Repeat source
                 baseBuf = startBuf;
@@ -342,6 +342,6 @@ namespace Ogre
     //-----------------------------------------------------------------------
     unsigned short InstanceBatchShader::getNumWorldTransforms(void) const
     {
-        return mNumWorldMatrices;
+        return uint16(mNumWorldMatrices);
     }
 }
